# .static/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing wallets with deposit and withdrawal operations",
        "title": "Wallet Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/wallet": {
            "post": {
                "description": "Executes a deposit or withdrawal operation on a wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Execute wallet operation",
                "parameters": [
                    {
                        "description": "Operation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.OperationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or insufficient funds",
                        "schema": {
                            "$ref": "#/definitions/ITK_pkg_api_response.Response"
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "schema": {
                            "$ref": "#/definitions/ITK_pkg_api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ITK_pkg_api_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/wallet/create": {
            "post": {
                "description": "Creates a new wallet with zero balance and returns its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Create new wallet",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.CreateWalletResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ITK_pkg_api_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/wallets/{id}": {
            "get": {
                "description": "Returns the current balance of a wallet",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Get wallet balance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.BalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid wallet ID",
                        "schema": {
                            "$ref": "#/definitions/ITK_pkg_api_response.Response"
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "schema": {
                            "$ref": "#/definitions/ITK_pkg_api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ITK_pkg_api_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ITK_pkg_api_response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.BalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number",
                    "example": 5000.5
                },
                "walletId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "internal_api_handlers.CreateWalletResponse": {
            "type": "object",
            "properties": {
                "walletId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "internal_api_handlers.OperationRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 1000.5
                },
                "operationType": {
                    "type": "string",
                    "enum": [
                        "DEPOSIT",
                        "WITHDRAW"
                    ],
                    "example": "DEPOSIT"
                },
                "walletId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "internal_api_handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}